+++ src/Views/TitleBar.react.js
@@ -1,5 +1,6 @@
 'use strict';
 
+import _ from 'lodash';
 import ClassNames from 'classnames';
 import MediaQuery from 'react-responsive';
 import PropTypes from 'prop-types';
@@ -8,37 +9,45 @@ import React, { Component } from 'react';
 import Header from '../Elements/Header.react';
 
 class TitleBar extends Component {
-
     render() {
-        const containerClasses = ClassNames('ui', 'title-bar', this.props.className);
+        const { children, className, style, subTitle, title } = this.props;
+        const containerClasses = ClassNames('ui', 'title-bar', className);
 
         return (
-            <header className={containerClasses} style={this.props.style}>
+            <header className={containerClasses} style={style}>
                 <MediaQuery maxWidth={767}>
                     {matches => {
                         return (
                             <Header
-                                as={matches ? 'h2' : 'h1'}
-                                style={{
-                                    overflow: 'hidden',
-                                    textOverflow: 'ellipsis',
-                                    whiteSpace: 'nowrap'
-                                }}
+                                as={matches ? 'h4' : 'h2'}
+                                sub={!_.isEmpty(subTitle)}
                             >
-                                {this.props.title}
+                                <span>{title}</span>
+
+                                {!_.isEmpty(subTitle) ? (
+                                    <Header.Subheader>
+                                        {subTitle}
+                                    </Header.Subheader>
+                                ) : null}
                             </Header>
                         );
                     }}
                 </MediaQuery>
+
+                {children ? (
+                    <div className="title-bar-children">
+                        {children}
+                    </div>
+                ) : null}
             </header>
         );
     }
-
 };
 
 TitleBar.propTypes = {
     className: PropTypes.string,
     style: PropTypes.object,
+    subTitle: PropTypes.string,
     title: PropTypes.string
 };